requirements.txt
requests~=2.21.0
xmltodict~=0.13.0
pandas~=1.5.2
SQLAlchemy~=2.0.21
python-dotenv~=1.0.0
matplotlib~=3.6.2
scipy~=1.9.3